pyproject.toml
[tool.poetry]
name = "fs.onedrivefs"
packages = [
    { include = "fs"}
]
version = "0.3.0"
description = "Pyfilesystem2 implementation for OneDrive using Microsoft Graph API"
authors = ["Rehan Khwaja <rehan@khwaja.name>"]
license = "MIT"
homepage = "https://github.com/rkhwaja/fs.onedrivefs"
keywords = ["filesystem", "Pyfilesystem2", "onedrive"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.6"
# Need 2.0.5 for opendir factory parameter
# Need 2.0.6 because in 2.0.5, opener didn't work
fs = ">=2.0.6"
requests = "^2.20"
requests-oauthlib = "^1.0"
pynacl = "^1.4.0"

[tool.poetry.dev-dependencies]
pyperclip = "^1.7"
pytest = ">=3.10"
pylint = ">=2.5.3"
pylint-quotes = "^0.2.1"
pytest-localserver = "^0.5.0"
pyngrok = "^1.4"
click = "^7.0"
pytest-cov = ">=2.9.0"

[tool.poetry.plugins] # Optional super table

[tool.poetry.plugins."fs.opener"]
"onedrive" = "fs.onedrivefs.opener:OneDriveFSOpener"

[tool.pylint.master]
load-plugins = "pylint_quotes"

[tool.pylint.format]
indent-string = "\t"
max-line-length = "200"
string-quote = "single-avoid-escape"

[tool.pylint.messages_control]
# no-member to remove false positives with objects constructed via google's discovery protocol
disable = [
	"bad-continuation",
	"duplicate-code",
	"fixme",
	"invalid-name",
	"len-as-condition",
	"logging-format-interpolation",
	"logging-fstring-interpolation",
	"missing-docstring",
	"relative-beyond-top-level",
	"too-many-locals"
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
